+++ CHANGELOG.md
@@ -11,7 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 
 ### Changed
 
-- SDK: Logout clients on profile deletion
+- SDK: Logout clients on active profile deletion
 
 ### Deprecated